requirements.txt
certifi
requests
praw
PyYAML
